Decompile renderCutsceneEditorText

+++ include/cutscene/cutscene_editor_helpers.h
@@ -5,8 +5,8 @@
 
 typedef struct CutsceneEditorTextGrid CutsceneEditorTextGrid;
 
-void renderCutsceneSlotMenuItem(CutsceneEditorTextGrid *grid, s16 row, s16 colorIndex);
-void renderCutsceneEditorText(s32 uiResourceId, s32 x, s32 y, char *text, s32 colorIndex);
+void renderCutsceneSlotMenuItem(CutsceneEditorTextGrid *grid, s16 row, u16 colorIndex);
+void renderCutsceneEditorText(s32 uiResourceId, s32 x, s32 y, char *text, u16 colorIndex);
 void setupCutsceneCommandLayout(s32 commandIndex, s32 x, s32 y, s32 width, s32 height, s32 flags);
 
 #endif
